deployment-arbitrary-cluster.md: Fix instructions. (#146)
* deployment-arbitrary-cluster.md: Fix instructions.

The sample commands quoted the list of filenames, which makes them
a single item.  The whole list of filenames is then used in a single kubectl
command.  Just drop the quotes so each filename is handled by itself.

Also change kubectl to oc for the OpenShift instructions.

* deployment-arbitrary-cluster.md: Fix manifest paths.

The crds/ and openshift/ prefixes should not be used when downloading
these manifests from the releases page.  They are available at the
same level as the other manifests.

* deployment-arbitrary-cluster.md: Add missing CRD.

Add the missing NodeNetworkConfigurationPolicy CRD to the
instructions.  Otherwise the nmstate-handler will crash loop.

* deployment-arbitrary-cluster.md: Add namespace.