Support untagged traffic at linux bridges
We were configuring both the host interface and the bridge to allow all vlans
but we also remove the "1 PVID Egress Untagged" from both the bridge and host interface,
this mean untag traffic (all the traffic from the bridge interface ip) will not be able
to exit the bridge interface because it doesn't have tag. This PR just change vlan filtering
for uplink interfaces and not for the bridge itself.