Test that connections persist a reboot
It will create a vlan interface with a policy and reboot the nodes
preventing k8s from starting up, the vlan connection has to be there
since is NM the one configured and it's persistent.

The teardown code puts back the k8s configuration and ensure that
k8s is at good shape.

Closes #143

Signed-off-by: Quique Llorente <ellorent@redhat.com>