Build and deploy using travis on merge
This PR add a deploy section to the travis yaml.
On every PR that is merge to master we are going to build and push
the nmstate-handler container to quay.io with laster tag.

when we create a git tag we will release a container with the same
tag on quay

Signed-off-by: Sebastian Sch <sebassch@gmail.com>