NodeSelector change does not trigger reconcile
When `nodeSelector` is changed and it matches a new node, this node ignores it.

This issue was caused by reconcile predicate `UpdateFunc` callback depending on
matching the node with both the new and the old policy `nodeSelector`. This
caused `UpdateFunc` to not work properly.

Now we check our node only against the new (updated) policy `nodeSelector`.

Signed-off-by: HlinaCZ <djhlinacz@gmail.com>