add list of conditions to policy
This list will be used to keep configuration status per each node
in Policy's State.

That will be a temporary solution until we introduce Enactment object.

The benefit of this is, that we keep the status of configuration in
the object that is used to trigger the configuration. In other words,
we decouple reporting and configuration.

Signed-off-by: Petr Horacek <phoracek@redhat.com>