Decouple configuration and reporting (#262)
* pass raw State to helper functions (#249)

Make set functions more generic by changing their input type to
plain State (thus not coupled with NNS).

* remove Spec from NodeNetworkState (#251)

The NNS will become read-only and desired state will be moved to
NNCP.

* drop NodeNetworkState config controller (#252)

With configuration being moved to the NNCP controller, there is no need to keep
the NNS config controller anymore. Therefore, this patch is removing it.

This patch also renames NNS report controller to simply NNS controller since
there is no need to differentiate between two anymore

* add list of conditions to policy (#247)

* add list of conditions to policy

This list will be used to keep configuration status per each node
in Policy's State.

That will be a temporary solution until we introduce Enactment object.

The benefit of this is, that we keep the status of configuration in
the object that is used to trigger the configuration. In other words,
we decouple reporting and configuration.

* add policy condition list helpers

These helpers will be used to A) set configuration condition for a specific node
and B) to obtain condition of certain type for a given node.

* we forgot to drop this as a part of #251 (#253)

Move configuration code from nns to nncp

add configuration code to policy controller (#256)

Run `nmstatectl set` directly from the NNCP controller. Tests will be adjusted
in follow-up PR since the scale is just too big to easily review in a single PR.

* Rebase decouple configuration and reporting (#260)

* Support ocp provider (#254)

For ocp we need the scc manifests and the
network interface names are as oks

* Use go modules and operator sdk 0.12.0 (#246)

For tools vendoring we are using [1] mechanism.

[1] https://github.com/golang/go/issues/25922

* Remove knmstate from vendor dir (#258)

* Use NNCP to set desiredState at e2e tests (#261)

Now that we have move desiredState to NNCP we have to change
e2e to use it instead of NNS to fix them.

* Update CRD with changed policy

Signed-off-by: Petr Horacek <phoracek@redhat.com>
Co-authored-by: Quique Llorente <ellorent@redhat.com>