WIP: perform dry run before applying configuration
In order to support multiple policies changing the same node, we plan
to periodically reapply all matching policies on the node. In order not
to run the configuration every time we re-reconcile, we have to perform
a dry run first. Since the dry run is not yet supported in nmstate, we
introduce a script which uses libnmstate functions to perform the check.

Signed-off-by: Petr Horacek <phoracek@redhat.com>