drop outdated guides and merge developer guide
Our design documents are outdated. It is hard to keep them up to date
while the project is evolving so fast. For now, let's just drop them
and replace them by a proper design document later.

This patch also merges developer-guide-commands.md into the CONTRIBUTION
document in order to keep all the needed info in a single place and hopefully
make it more visible and easier to keep up to date.

Signed-off-by: Petr Horacek <phoracek@redhat.com>