Add Alona, Eran and Igal to reviewers (#372)
Alona, Igal And Eran contributed to the project and deserve to be recognized as reviewers. By adding them to the list, they will gain access to `/lgtm` command and therefore will be able to confirm merging of PRs. They will be also assign by our not to review new PRs.

Signed-off-by: Petr Horacek <phoracek@redhat.com>