explicitly bring unavailable interfaces up (#423)
There is a bug in Kernel/NetworkManager on systems with NetworkManager
1.20, where sometimes after disconnecting a NIC from a bonding, the NIC
remains in 'unavailable' state and cannot be used for a new connection. This
is likely caused by an issue with autonegotiation where the NIC appears to
be disconnected and the only thing that can bring it available again is
explicitly calling ip link set up on it. In order to workaround
this issue until it gets solved, we iterate all devices during nmstatectl set
and if we find some with 'unavailable' we explicitly set them up.

Signed-off-by: Petr Horacek <phoracek@redhat.com>