Move handler tests to new target
In order to run the handler e2e on their own, this moves them to their
own Makefile target, but maintains the old e2e target by using the new
one as a dependency. Then, when operator e2e tests are introduced, they
can both be called individually or as a group via test/e2e.

Signed-off-by: Brad P. Crochet <brad@redhat.com>