drop metrics from handler (#518)
kubernetes-nmstate handler has to run in host network namespace. Due to
that, it shares the pool of available TCP ports with other services
running on the host. This is dangerous, since if there is a collision,
kubernetes-nmstate will not be able to start and there is no way to
change the port setting.

Since we don't really test metrics (and I'm not aware of anybody using metrics
with kubernetes-nmstate), this patch is dropping metrics from the
project. With that, we should avoid any port conflicts and also save
some resource requirements.

Signed-off-by: Petr Horacek <phoracek@redhat.com>