Move handler tests to new target (#517)
* Move handler tests to new target

In order to run the handler e2e on their own, this moves them to their
own Makefile target, but maintains the old e2e target by using the new
one as a dependency. Then, when operator e2e tests are introduced, they
can both be called individually or as a group via test/e2e.

Signed-off-by: Brad P. Crochet <brad@redhat.com>

* Add operator suite of e2e tests

There are currently no test in it, but the framework is there.

Signed-off-by: Brad P. Crochet <brad@redhat.com>

* Separate handler and operator test scripts

Leave the existing scripts in automation/ as handler e2e tests, and
add scripts for the operator e2e suite. Once these land, add them
to prow.

Signed-off-by: Brad P. Crochet <brad@redhat.com>

* Run the operator locally for operator e2e tests

Signed-off-by: Brad P. Crochet <brad@redhat.com>