Add optional node selector field to NMstate CRD (#514)
In current implementation the handler pods deployed on all the masters and workers nodes.
With node selector the user can specify on which nodes the handler should run. in order to do that
,the user should first set the label on the relevant nodes and then set the same label value in NMstate node selector field.

This PR also updates controller to return erros if it didn't manage to render at least one file for each manifest directory.

Signed-off-by: Yossi Boaron yboaron@redhat.com

Co-authored-by: Yossi Boaron <yossi.boaron.1234@gmail.com>