Makefile: remove build dependency on generated files (#533)
If generated files are not checked in when they should be, a diff
between a built image and what is in the repo could occur. This would
not be desirable. So, remove the Makefile dependency, and add a check
script that makes sure that what is commited to the repo is the same
as what should be committed.

Signed-off-by: Brad P. Crochet <brad@redhat.com>