Bump operator-sdk-0.18.0 (#564)
* Bump operator-sdk-1.18.0

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Regenerate manifests after bumping operator-sdk to 0.18.0

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Adapt tests to new operator-sdk version

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Activate back CRD's preserve-unknown-fields needed by State

Bumping to this operator-sdk versions means that CRD k8s type bumps to v1, this version
by default deactivate `preserve-unknonw-fields` property [1] so it prunes our custom `State` field
since it's a json object with unknown fields that we just pass to nmstate.

To re-actativate it there is a kubebuilder flag to do so [2] so we can specify it at
`CurrentState` and `DesiredState` fields.

[1] https://github.com/kubernetes-sigs/controller-tools/blob/master/pkg/crd/gen.go#L54
[2] https://kubernetes.io/docs/tasks/access-kubernetes-api/custom-resources/custom-resource-definitions/#controlling-pruning

Signed-off-by: Quique Llorente <ellorent@redhat.com>