Test active-backup bonding with primary interface (#586)
Our functional tests of bonding configured as the default interface are
flaky. It may be due to the fact that sometimes we select different link
as the primary one, therefore get a different MAC, resulting in
different IP and broken node connectivity.

With this patch we switch the used mode to active-backup and explicitly
set the first NIC as the primary one. That should make sure that we
always start with the same active NIC and thus the same IP.

Signed-off-by: Petr Horacek <phoracek@redhat.com>