Reconcile populted NNCP if node is created or modified
Current kubernetes-nmstate version does not Reconcile already populated NNCPs
if Nodes are added to the cluster or node labels are modified. This PR force
Reconcile of all policies if a node is create or updated.

Since Reconcile functions already take care to checking node labels we choose
to simplify the implementation here just checking for Node Update not checking
if the update is for labels or if the labels are for this node, nmstate is declarative
so if if same version is re-apply it should not be a noop.

Signed-off-by: Quique Llorente <ellorent@redhat.com>