Remove code directories from .dockerignore (#633)
Remove pkg cmd and vendor from .dockerignore, as it causes builds
with Docker and the build/Dockerfile.*openshift builds to fail.
This is due to them being multistage builds.

The error is:

Step 4/10 : RUN go build --mod=vendor -o build/_output/bin/manager main.go
---> Running in 946711c1028e
go: inconsistent vendoring in /go/src/github.com/openshift/kubernetes-nmstate:

This is caused by the vendor directory not being present. This does not affect
Podman builds.

The multistage builds need to have a complete picture, so we need
to include these directories in the build context. This will allow the
regular Dockerfiles to also be migrated to a multistage build approach,
which is desirable.

Signed-off-by: Brad P. Crochet <brad@redhat.com>