Fix intermitent Status at NNCP sequencial rollout
Reconcile always resets nncp status at the beginning.
This doesn't cause issue when applied in parallel, but
in case of sequential rollout, the nncp is missing status
information for the whole duration of a reconcile.

Fixing this by resetting only when webhook is not used.
Without using webhook, nncp status may not always be displayed
while nncp is progressing. This does not affect functionality.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>