Reduce default gw probe traces (#645)
Currectly kubernetes-nmstate prints the network state every time it does
not find the default gw when the probe is running, clearly this is too
verbose. To fix that this change in case of "-v=production" just print
the error and the gjson path used to find the default gw, in case of
"-v=debug" it also prints the network state.

Signed-off-by: Quique Llorente <ellorent@redhat.com>