ci, Fix ci commands (#646)
* ci, Change ci Makefile commands use KUBECONFIG and OPERATOR_NAMESPACE

Currently test-e2e-handler and test-e2e-operator use explicit kubeconfig of the
knmstate repo. In order to run tier1 tests, we want to be able to use KUBECONFIG
set by external cluster.
We also add OPERATOR_NAMESPACE to ecplicitly mention that it is used in the tests.
Added OPERATOR_NAMESPACE var to makefile targets commands
Added KUBECONFIG var instead of explicit kubeconfig in ci makefile targets

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* ci, Remove usage of namespace in cluster scoped objects

Currently we are using the namespace to find NodeNetworkState object
This is wrong, since this object is cluster scoped.
Removed the usage of the namespac when searching of NodeNetworkState objects

Signed-off-by: Ram Lavi <ralavi@redhat.com>