tests: Show NNCE if NNCP is not matching (#650)
* tests: Show NNCE if NNCP is not matching

If a NNCP is not matching in the e2e tests the Reason and the Message is
missing from the output and also the NNCEs are not presented. This
change uses gomega gstruct matcher so we can have fuzzy match on Reason
and Message and still it will show it them and also add the NNCEs in
case of not matching.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* tests: Initialize correctly testenv

The OperatorNamespace variable was not initialized before kubernetes
reporter is created so it couldn't find pods to get the logs from.

Signed-off-by: Quique Llorente <ellorent@redhat.com>