Don't dump k8s pod logs at test output
After fix the k8s reporter a part from dumping pod logs to a file it
also dump it to GinkgoWriter, the GinkgoWriter show the content written
with it when the test fail. Dumping this to logs makes the output too
long and we already have it at the file, this change remove that and
make the reporter just dump to a file.

Signed-off-by: Quique Llorente <ellorent@redhat.com>