Add sleep after changing node labels at tests
The NNCPs are reconciled if node labels are changed, but there is no
webhook that blocks the operation until the NNCP status is reset, that's
make some times fall into a wrong assumptions around status. This change
add some sleep time after changing labels with a TODO about removing it
when webhook is implemented.

Signed-off-by: Quique Llorente <ellorent@redhat.com>