Bump nmstate to 0.3 (#643)
* Bump nmstate to 0.3

Looks like nmstate 0.2 is going to approach EOL so we bump to next minor
it also depends on NM 1.26. We are not going to bump kubecirtci node NM
since we have to check that this is retrocompatible.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Don't dump k8s pod logs at test output

After fix the k8s reporter a part from dumping pod logs to a file it
also dump it to GinkgoWriter, the GinkgoWriter show the content written
with it when the test fail. Dumping this to logs makes the output too
long and we already have it at the file, this change remove that and
make the reporter just dump to a file.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Print generated bridge1 and bond1 names

To reduce the collision between test in case tear down is broken we
generate a different name for bridge1 and bond1 variable per test, for
debugging purposes is good to know what this names are per test so you
can see at what test they were generated.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Install tools only at go.mod changes

After some changes at master the go install command to install the tools
is run everytime a make target that need the tools is executed, this
change leave as it was, creating a dependency between the binary and
go.mod.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Adapt thest to nmstate-0.3 checks

After bumping to nmstate-0.3 a lot o test were failing with good reason,
they have bugs at tear-down code and some bridge/bonds where not
propertly deleted, this make nmstate-0.3 fail when another bridge/bond
is trying to use as slave a nic that already belongs to another
bridge/bond. To fix all this now the tests are very strict too and the
network configuration state check is done at all nodes not only at
workers since some of our tests are run at master nodes too (examples,
node label).

Also there is a bug a trace that prints a big and ugly error log related of using
WithValue instead of with name WithName at allPolicies function.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Add sleep after changing node labels at tests

The NNCPs are reconciled if node labels are changed, but there is no
webhook that blocks the operation until the NNCP status is reset, that's
make some times fall into a wrong assumptions around status. This change
add some sleep time after changing labels with a TODO about removing it
when webhook is implemented.

Signed-off-by: Quique Llorente <ellorent@redhat.com>