Remove gc-timer at e2e tests when comparing NNS currentState (#668)
* Print NNS test currentState as string

Comparin directly the NNS Status struct print currentState as slice of
bytes that does not help to find differences, this change tranform the
obtained and expected to strings to make easier to find differences.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Remove gc-timer from e2e tests check

When NNS current is compared the dynamic attributes has to be removed
for it to match like is already dont at controller.

Signed-off-by: Quique Llorente <ellorent@redhat.com>