Generate webhooks with operator-sdk and add logic
The code from kube-admission-webhook library used to setup webhooks is
very minimal and only make sense if the pod where is running the webhook
also runs the cert manager, by moving the cert-manager to a different
pod we can use just the cert-manager part from kube-admission-webhook
and use kubebuilder style webhooks.

The webhook logic has being simplified and the NNCP Status conditions
are just removed but not set to unknown since changing a subresource is
not easy with kubebuilder way of doing webhooks.

Signed-off-by: Quique Llorente <ellorent@redhat.com>