state, filter: Treat numeric-only iface names
When processing the filtering of the interfaces status, the interfaces
with names that include only numeric characters are interpreted as
float64.

In order to (partially) support such interfaces, make sure to detect them
and avoid their filtering. I.E. these interfaces cannot be filtered-out.

Note: This solution is partial due to the improper interpretation of
special numeric (only) names such as (123.0 and 0xff). These will
wrongly get interpreted as 123 and 255.

Signed-off-by: Edward Haas <edwardh@redhat.com>