Webhookconfig v1 (#712)
* Bump kaw v0.14.0 and operator-sdk v1.4.2

The webhookconfiguration v1alpha1 is going to be deprecated at
kubernetes 1.22. This change bumps operator-sdk and
kube-admission-webhook to support webhookconfiguration v1

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Adapt code to new operator-sdk and kaw versions

After bumping operator-sdk to 1.4.2 the Reconcile function expects a
context.Context and the Meta at events is included in the client.Object
passed to it.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Don't use .gitignore at automation rsync

After bumping deps some of the .gitignore in the project is making rsync ignore
"vendor/github.com/moby/term/.gitignore" and this breaks "make
check-gen"

Signed-off-by: Quique Llorente <ellorent@redhat.com>