Use github.com/gofrs/flock to lock handler (#731)
* Use github.com/gofrs/flock to lock handler

The nmstate-handler has a lock that make it wait if another handler is
already running in the system, but the library used for that does not
work well with containers. This change replace that library with
different one that uses golang syscall.Flock behind the scine [1] and
that works fine with containers sharing node volume.

[1] github.com/nightlyone/lockfile.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Replace handler readiness probe

The nmstate handler is marked as ready even if it was not able to take
ownership of the lock, this change runs nmstatectl and touchs a file
after gaining lock ownership so the readiness probe can check both
things.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Test nmstate handler lock mechanism

The nmstate handler don't get ready if another handler is running at the
same node. This change add an e2e tests to check that this is working.

Signed-off-by: Quique Llorente <ellorent@redhat.com>