Reorder nnpc controller functions
The first thing that reader is presumably interested in when opening
nodenetworkconfigurationpolicy_controller.go file is the Reconcile loop.
In this commit, I'm moving the Reconcile loop to the top of the file,
and reordering the helper functions so that the reader can see the
details in the same order in which these helper functions are called in
Reconcile.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>