cluster-sync, Fix Eventually and consistently logic
Current logic of Eventually and consistently
would not trigger the last needed iteration,
since it compare the count to zero after sleeping.
Fix the logic, by doing the comparison in the beginning
of the loop, this way it won't sleep on the last unneeded
iteration, and will also perform that last iteration.

Signed-off-by: Or Shoval <oshoval@redhat.com>