Bump nmstate to 1.0 (#745)
* Bump nmstate to 1.0

Also PIN future to nmstate-git and NetworkManager-main copr repos.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Increase e2e handler test timeout to 2h

Looks like it's taking more time to complete now that we have bump
nmstaet and NM.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Remove references to slavery terms

There are some parts of the tests that use the old nmstate schema where
ports are named "slaves". This change remove that now that we are at
nmstate > 1.y.z.

Signed-off-by: Quique Llorente <ellorent@redhat.com>