Add "Pending" condition to nnce status (#748)
* Add "Pending" condition to nnce status

After removing the unmatching enactments we don't have the
pre-progressing state. This change add a state that marks the nnce as
Pending when max unavailable node limit is reached and the node has to
wait for other to finish.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Use condition.type as the status column

The reson is too long for the column and saying just "Pending" at
MaxUnsavailableLimitReach is enough.

Signed-off-by: Quique Llorente <ellorent@redhat.com>

* Adapt docs to STATUS changes

The STATUS column now shows condition.Type and not condition.Reason.
This change adapt docs to that.

Signed-off-by: Quique Llorente <ellorent@redhat.com>