Relationship labels, Label self deployed resources (#770)
In order to have relationship (app.kubernetes.io/*)
on all deployed resources,
this commit handles the resources that are deployed by itself
such as:
1. Node Network Configuration Enactments
2. Node Network State

The labels are inherited from the pod, if exists.
The pod will have those labels if deployed by OLM / HCO,
otherwise it won't and the labels won't be added to the above
resources.

Signed-off-by: Or Shoval <oshoval@redhat.com>