[release-0.47] Avoid incrementing NRU when NNCE
At the beginning, check if NNCE exists for a policy and is in
progressing state.

If it is, then it means that handler was killed while applying the
desired state - do not increment unavailableNodeCount in such case.

There is still a window where killing handler would cause
inconsistency - between status is set to Failed or Success and the
unavailableNodeCount is decremented.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>
Signed-off-by: Quique Llorente <ellorent@redhat.com>