Use go cmd from go get
The install-go.sh install the golang version from go.mod under ./build
directory to build the project with a specific version. This is no
longer ncessary since go allows to download a specific version of golang
with "go get golang.org/dl/go1.10.7" and then "go download go1.10.7",
the only issue is it requieres a go to be installed.

Signed-off-by: Quique Llorente <ellorent@redhat.com>