Use custom Backoff retries setting for status update (#786)
Applying a policy that breaks node connectivity causes client
to return cached NNCP resource with outdated resourceVersion
until watchers reconnect again, which takes something around
half a minute.

Using outdated policy leads to conflicts at the end of Reconcile,
where the Policy status is recalculated.

This commit adds custom StatusUpdateRetry Backoff that increases
number of attempts and factor to allow watchers to reconnect.

This custom StatusUpdateRetry is used in functions, that are called
after the policy desiredState was applied (and potentionally broke the
node) - decrementUnavailableNodeCount and policyconditions.Update
functions.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>