Squash nns controller into node controller
The reconcile loops of both of the controllers were pracitically the same.
The only difference was the NodeController stored the previous state and
updated the NNS only if the state was changed.
When squashing, this logic had to be changed since removal of an exisiting
NNS may be skipped. Now the nns is updated/re-created if the nns doesn't
exist or the state was changed.

The main changes to the exsisting flow are-
1. The existance of NNS will be verified anyway (so in the common use case
of periodic update that happens every minute, there will be one extra,
probably redundant, api call. I beleive it worth the duplication removal).
2. If the state was not changed since the last Reconcile, even in case of
force-update, the NNS won't be updated.


Signed-off-by: Alona Kaplan <alkaplan@redhat.com>