e2e: Add test verifying interface_filter isn't failing
Since NetworkManager client could fail to connect to dbus
(potentially), we should make sure that we catch such
events.

When the NetworkManager client fails to connect, the NNS
is not filtered and error message is logged by handler.

This commit adds test verifying that handler log doesn't
contain such message.

Signed-off-by: Radim Hrazdil <rhrazdil@redhat.com>